refactor(core): move code to better named src/ folder
