+++ core/CMakeLists.txt
@@ -13,5 +13,5 @@ set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
 
-# Build host first to ensure logos_core is built before modules
-add_subdirectory(host)
+# Build src first to ensure logos_core is built before modules
+add_subdirectory(src)
