+++ core/src/core_manager.cpp
