+++ core/src/core_manager.h
