+++ core/src/core_manager_interface.h
