+++ core/src/logos_core.cpp
