+++ core/src/logos_core.h
