+++ core/src/metadata.json
