+++ examples/custom_app/custom_app.cpp
@@ -3,7 +3,7 @@
 #include <QString>
 #include <QTimer>
 #include <QCoreApplication>
-#include "../../core/host/logos_core.h"
+#include "../../core/src/logos_core.h"
 
 // Helper function to print loaded plugins
 void printLoadedPlugins() {
