+++ examples/standalone_chat/CMakeLists.txt
@@ -17,7 +17,7 @@ endif()
 # Include directories for the project
 include_directories(
     ${CMAKE_CURRENT_SOURCE_DIR}/../../core
-    ${CMAKE_CURRENT_SOURCE_DIR}/../../core/host
+    ${CMAKE_CURRENT_SOURCE_DIR}/../../core/src
     ${CMAKE_CURRENT_SOURCE_DIR}/../../logos_app/app/interfaces
 )
